feature: add abstract interface for llama and chatglm * TODO: implement the chatglm2 train
