+++ aspen/__init__.py
@@ -1,16 +1,18 @@
 from aspen.utils import convert_hf_to_pth, save_lora_model
 from aspen.tokenizer import Tokenizer
-from aspen.model import LlamaModel, Linear, RMSNorm
-from aspen.modelargs import TokenizerArgs, LlamaModelArgs, MultiLoraBatchData, LoraBatchDataConfig
+from aspen.model import LLMModel
+from aspen.model_llama import LlamaModel
+from aspen.model_chatglm import ChatGLMModel
+from aspen.modelargs import TokenizerArgs, LLMModelArgs, MultiLoraBatchData, LoraBatchDataConfig
 from aspen.dispatcher import TrainTask, Dispatcher
 
 __all__ = [
     "Tokenizer",
+    "LLMModel",
     "LlamaModel",
-    "Linear",
-    "RMSNorm",
+    "ChatGLMModel",
     "TokenizerArgs",
-    "LlamaModelArgs",
+    "LLMModelArgs",
     "MultiLoraBatchData",
     "LoraBatchDataConfig",
     "convert_hf_to_pth",
