+++ aspen/modelargs.py
@@ -14,7 +14,7 @@ class TokenizerArgs:
 
 
 @dataclass
-class LlamaModelArgs:
+class LLMModelArgs:
     dim_: int = 4096
     multiple_of_: int = 256
     n_heads_: int = 32
